app/backend/requirements.txt
openai==0.27.8
Flask==2.2.5
azure-identity==1.13.0
azure-search-documents==11.4.0b3
azure-storage-blob==12.16.0
langchain==0.0.225
pypdf==3.12.0
azure-ai-ml==1.8.0
azure-ai-formrecognizer==3.2.1
docx==0.2.4
docxtpl==0.11.2